Clean up with pixel and Region imheritance
Remove pixel base struct.
Move Region implementation to header.
Make Region as a template.
Create a RegionBase.

+++ include/config.h
@@ -3,7 +3,7 @@
 
 #include <cstdint>
 
-extern uint32_t jacardThreshold;
+extern double jacardThreshold;
 extern uint32_t similarityThreshold;
 
 #endif /* _CONFIG_H_ */
